Began refactor of xyz loading
